chore(src/components/inputs/styles): move to packages
