+++ packages/components/inputs/number-input/src/number-input.js
@@ -4,7 +4,7 @@ import invariant from 'tiny-invariant';
 import requiredIf from 'react-required-if';
 import { filterDataAttributes } from '@commercetools-uikit/utils';
 import Constraints from '@commercetools-uikit/constraints';
-import { getInputStyles } from '../../../../../src/components/inputs/styles';
+import { getInputStyles } from '../../styles';
 
 const NumberInput = (props) => (
   <Constraints.Horizontal constraint={props.horizontalConstraint}>
